boost169: adaption of the code to 1.69

+++ src/xmpp/xmpp_config.h
@@ -30,7 +30,7 @@ public:
     std::string path_to_ca_cert;
     int tcp_hold_time;
     bool gr_helper_disable;
-    uint8_t xmpp_hold_time;
+    int xmpp_hold_time; // was uint8_t, but int everywhere
     uint8_t dscp_value;
     std::string xmlns;
